moving a few changes to the right place after rebase
