+++ extensions/dacpac/src/controllers/mainController.ts
@@ -6,21 +6,17 @@
 'use strict';
 
 import * as azdata from 'azdata';
+import ControllerBase from './controllerBase';
 import * as vscode from 'vscode';
 import { DataTierApplicationWizard } from '../wizard/dataTierApplicationWizard';
 
 /**
  * The main controller class that initializes the extension
  */
-export default class MainController implements vscode.Disposable  {
-	protected _context: vscode.ExtensionContext;
+export default class MainController extends ControllerBase {
 
 	public constructor(context: vscode.ExtensionContext) {
-		this._context = context;
-	}
-
-	public get extensionContext(): vscode.ExtensionContext {
-		return this._context;
+		super(context);
 	}
 	/**
 	 */
@@ -34,9 +30,5 @@ export default class MainController implements vscode.Disposable  {
 
 	private initializeDacFxWizard() {
 		azdata.tasks.registerTask('dacFx.start', (profile: azdata.IConnectionProfile, ...args: any[]) => new DataTierApplicationWizard().start(profile, args));
-	}
-
-	public dispose(): void {
-		this.deactivate();
 	}
 }
