+++ extensions/dacpac/src/main.ts
@@ -5,9 +5,10 @@
 'use strict';
 import * as vscode from 'vscode';
 
+import ControllerBase from './controllers/controllerBase';
 import MainController from './controllers/mainController';
 
-let controllers: MainController[] = [];
+let controllers: ControllerBase[] = [];
 
 export function activate(context: vscode.ExtensionContext) {
 	let activations: Promise<boolean>[] = [];
