+++ extensions/schema-compare/src/main.ts
@@ -5,10 +5,9 @@
 'use strict';
 import * as vscode from 'vscode';
 
-import ControllerBase from './controllers/controllerBase';
 import MainController from './controllers/mainController';
 
-let controllers: ControllerBase[] = [];
+let controllers: MainController[] = [];
 
 export function activate(context: vscode.ExtensionContext) {
 	let activations: Promise<boolean>[] = [];
